Make search connection pool configurable
